feat: lint commit messages
Add a pre-commit hook to lint commit messages, following the
conventional commit style [1].

The Azure Pipelines is also extended with the same linter, but in
its original JavaScript form to lint all commits since master. This
ensures that a merge request with several commits will be thoroughly
validated with regards to commit messages.

[1]: https://www.conventionalcommits.org/en/v1.0.0/

Fixes: #582