fix: tweak the conventional commits changelog preset
Use the conventional commits preset for the changelog but still
tweak it to 1) not show links and 2) hide the commit short hash.
The majority of our codebase is proprietary (and thus required
authenticated access) and the commit hashes only add clutter
(when needed, git-log does a better job).

Manually fix the changelog.

Fixes: #586