feat: create container recipes to replace pre-commit steps
Create container recipes for the (code) formatters and (code) linters
(aggregated into "flinters"), as well as for the commit linters.

The Azure pipeline QA steps are adjusted to use container images instead
of trying to cache and re-install all the toolchains every time.

The container images must be created manually and pushed to the registry.

Fixes: #838