fix: yarn dependency cache should always depend on yarn.lock
The yarn dependency cache check would sometimes use the fallback restore
keys that did not depend on the yarn.lock file, meaning that in some
cases we would not invalidate node_modules when that behaviour was
desired. This commit removes the fallback restore keys: we should always
rebuild this cache if yarn.lock changes.

Fixes: #1603